Added JDBC based authentication
Replacing the in-memory test user, JDBC based authentication is used for
the application moving forward, including some customization to account
for the table names.

The datasource defined in the properties file is used, as well as the
BCrypt Password Encoder that's defined as the bean in the configuration
file.

By default, Spring Security expects users and authorities tables, but if
these names are different in the project, additional method calls are
required to provide the SQL needed to pull the authentication details.
These SQLs, based on the application's database design, is added to the
config file.