Added User List & other User-related functionalities
A User List view has been added to the application, where all registered
users are listed in a table, with Edit and Activate/Discontinue buttons
being present. Only users with 'ADMIN' role are able to access this page
from the navbar - the link appears accordingly.

This commit involves adding the appropriate handler methods to the
controller, along with addition of the User List view. Also, some of the
functions in the controller and service have been modified to account
for code flexibility.

Finally, the user view has been refactored to use it for both user
registration and modification. Also, message properties have been added
to the appropriate file to show the labels and prompts, as per the
requirement of the new feature and refactors.