Added Task List feature to the application
This commit also includes refactoring Task related controller and
services to fully enable the features of the application.

The handlers for the list page was added to the TaskController, along
with the service and validators required to perform the actions taken in
the page. The edit task page has been improved, with the action buttons
in it being made functional, along with the addition of delete/archive
action.