Implemented UserDetails to include extra params
The default implementation of UserDetails - in the form of User class -
only contains the username, enabled and the list of privileges. It was
needed for the application to have the user id, first name and last name
available as well.

This commit includes the custom implementation that includes the fields,
as well as include an override of UserDetailsService, to populate the
additional params into the object. The service implementation also
replaces the jdbcAuthentication that was used in the config, as well as
removal of the custom SQL queries (the userRepository handles the user
fetching now).