Added Pending Task list to Dashboard
This commit adds the Pending Task list to the dashboard, once the user
logs in to the application. The user can take actions, such as edit, mark
as complete and delete, from the said list.

In addition, the user's full name will appear on the navbar, instead of
the username, as it was shown before.