Added basic logging to application
Logging has been added to the controllers, with each of the handler
methods having at least one logging in them, ensuring that every flow is
logged at a debug level.

The application.properties file has been added to as well, in order to
include the property that only sets debug level logging for the project
files under 'com.zedapps' - not all the dependencies and such.