Removed standalone Tomcat deployment dependency
Previously, in order to run the application, one had to configure the
standalone Tomcat server. The application had to be built using gradle
and then had to be run using Tomcat, setting the context-path as 'todo'.

This dependency has since been removed in favor of using the embedded
server. The JSP parser library of Tomcat is added as a dependency to
ensure that Spring Boot's embedded server can parse the jsp files and
show the views properly.

In addition, the readme file has been updated to reflect the changes
made, along with fixing some typos.